% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.r
\name{Attach_QC}
\alias{Attach_QC}
\title{Attach_QC}
\usage{
Attach_QC(sc_obj, pattern = "default")
}
\arguments{
\item{sc_obj}{an sc_obj that containing a single cell RNA sequencing count matrix}

\item{pattern}{a regular expression or vector of regular expressions to select
the features for which outliers should be removed. The default pattern is
pattern = ("^MT-", "^RP").}
}
\value{
sc_obj
}
\description{
Automates quality control (QC) steps and attaches QC metrics to sc_obj
}
\examples{
data.path <- "path/to/file/"
counts <- import10x(data.path)
sce <- newSC_obj(counts)
sce <- Attach_QC(sce, pattern = "^MT")


}
