% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MULTIseq.r
\name{classify_cells}
\alias{classify_cells}
\title{classify_cells
Performs cell classification for a given quantile q}
\usage{
classify_cells(barTable, q)
}
\arguments{
\item{q}{}
}
\value{
Calls for quantile q
}
\description{
classify_cells
Performs cell classification for a given quantile q
}
