% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.r
\name{colnames.unique}
\alias{colnames.unique}
\title{colnames.unique
Takes a vector of meta data column names that contain duplicates
and returns a vector of unique column names}
\usage{
colnames.unique(char_names)
}
\arguments{
\item{char_names}{}
}
\value{
a vector of unique column names
}
\description{
This function relies on coercion of character values -> NA to detect
names that have already been made unique and insure that they are not further
modified.
}
\examples{
col.names <- colnames(metaData(sc_obj))
col.names <- colnames.unique(col.names)
}
