% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.r
\name{detect_anomalies}
\alias{detect_anomalies}
\title{detect_anomalies}
\usage{
detect_anomalies(sc_obj, method = "isolation", verbose = T)
}
\arguments{
\item{sc_obj}{an sc_obj with attached QC}

\item{method}{By default method is set to "isolation" for anomaly detection. "doublet_scoring" may be used
to perform troubleshooting tasks but will not detect anomalies.}

\item{verbose}{By default is set to True}
}
\value{
an sc_obj with identities of outlier events included in the metadata.
Visualization for anomaly detection is included in the sc_obj graphs.
}
\description{
Uses isolation forest to perform multivariate anomaly detection across all columns
present in sc_obj metadata
}
\examples{

  sce <- detect_anomalies(sce)

}
