% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_doublet_scoring.r
\name{est_doublets}
\alias{est_doublets}
\title{est_doublets
dispatches count matrix to scds to estimate doublet rate
refer to https://github.com/kostkalab/scds
doi: 10.1093/bioinformatics/btz698}
\usage{
est_doublets(counts, method = "hybrid")
}
\arguments{
\item{counts}{alternately accepts SingleCellExperiment Object}

\item{method}{defaults to "hybrid". Alternately "cxds" or "bcds" can be used}
}
\value{
estimate of the proportion of doublets
}
\description{
est_doublets
dispatches count matrix to scds to estimate doublet rate
refer to https://github.com/kostkalab/scds
doi: 10.1093/bioinformatics/btz698
}
\examples{
est_doublets allows you to chose any of the three methods used to score doublets
provided by Bais et al.

est_doublets(counts, method = "hybrid")

}
