% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTOobj.r
\name{generate_UMAP}
\alias{generate_UMAP}
\title{generate_UMAP
Performs dimension reduction on a formatted sample barcode table}
\usage{
generate_UMAP(n.bar.table, n_neighbors = 50, min_dist = 0.1)
}
\arguments{
\item{n_neighbors}{n neighbors parameter for the umap function from package uwot}

\item{min_dist}{minimum distance parameter for the umap function from package uwot}

\item{bar.table}{formatted sample barcode table}

\item{normalize}{logical, should the return table contain CLR normalized values. defaults to False.}
}
\value{
sample barcode table with UMAP dimensions UMAP1 & UMAP2 inserted as first two columns.
}
\description{
generate_UMAP
Performs dimension reduction on a formatted sample barcode table
}
\examples{
bar.UMAP <- generate_UMAP(bar.table, n_neighbors =50, min_dist = 0.1)
bar.UMAP <- generate_umap(bar.table)
}
