% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCobj.r
\name{Graph,SC_obj-method}
\alias{Graph,SC_obj-method}
\title{Graph}
\usage{
\S4method{Graph}{SC_obj}(x, index)
}
\arguments{
\item{x}{SC_obj.}
}
\value{
gridExtra output for stored graphs
}
\description{
Graph
}
\examples{
Graph(sc_obj, "Quality Check")
}
