% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTOobj.r
\name{HTO.format}
\alias{HTO.format}
\title{HTO.format}
\usage{
HTO.format(UMI_count_matrix)
}
\arguments{
\item{UMI_count_matrix}{raw count output}
}
\value{
formatted count matrix where each column corresponds to a barcode and each row to a cell UMI
}
\description{
Format function equipped to handle the outputs of CITE-seq count or the MULTIseq Alignment Suite
}
\details{
HTO.format() performs the following manipulations:
casts data input to a matrix using as.matrix()
transposes the input data if necessary
removes any uninformative columns attached during the alignment
}
\examples{
f.bar.table <- data.format(bar.table)
}
