% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTOobj.r
\docType{class}
\name{HTO_obj-class}
\alias{HTO_obj-class}
\alias{HTO_obj}
\title{HTO_obj}
\description{
HTO_obj
}
\section{Slots}{

\describe{
\item{\code{HTO.matrix}}{ANY.}

\item{\code{HTO.CLR_norm}}{ANY.}

\item{\code{UMAP}}{matrix.}

\item{\code{meta.data}}{data.frame.}

\item{\code{graphs}}{list.}
}}

