% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCobj.r
\name{HTOs,SC_obj-method}
\alias{HTOs,SC_obj-method}
\title{HTOs
retrieve HTO count table formatted for demultiplexing}
\usage{
\S4method{HTOs}{SC_obj}(x)
}
\arguments{
\item{SC_obj}{}
}
\value{
Table of Hashtag counts (barTable)
}
\description{
HTOs
retrieve HTO count table formatted for demultiplexing
}
