% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import10x.r
\name{import10x}
\alias{import10x}
\title{import10x
import 10x data from filtered feature matrices generated by Cell Ranger, and CITEseq count.
some public data sets are supported depending on file format and file naming convention.}
\usage{
import10x(path, type = "unspecified", ESNG = F, garbage_collection = T)
}
\arguments{
\item{path}{Accepts either a directory containing compressed 'tsv' files or an 'h5' file.}

\item{type}{(optional) Force input file type to user input. Supported input types are "ff_bc_matrix" or "h5".}

\item{ESNG}{Defaults to FALSE. Logical that determines weather gene names should be used or ensembl gene id}
}
\value{
Sparse matrix of 10x single-cell count data. Class dgCMatrix.
}
\description{
import10x
import 10x data from filtered feature matrices generated by Cell Ranger, and CITEseq count.
some public data sets are supported depending on file format and file naming convention.
}
\examples{
path_to_file <- '/path/to/data'
data <- import10x(path_to_file)

path_to_file <- '/path/to/data.h5'
data <- import10x(path = path_to_file, ESNG = T)
}
