% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importX.r
\name{importX}
\alias{importX}
\title{importX
import count matrices generated by Cell Ranger or CITEseq count.
some other count matrix formats are supported depending on
file format and file naming convention.}
\usage{
importX(path, type = "unspecified", ESNG = F, garbage_collection = T)
}
\arguments{
\item{path}{Accepts either a directory containing compressed 'tsv' files or an 'h5' file.}

\item{type}{(optional) Force input file type to user input. Supported input types are "ff_bc_matrix" or "h5".}

\item{ESNG}{Defaults to FALSE. Logical that determines weather gene names should be used or ensembl gene id}
}
\value{
Sparse matrix of single-cell count data. Class dgCMatrix.
}
\description{
importX
import count matrices generated by Cell Ranger or CITEseq count.
some other count matrix formats are supported depending on
file format and file naming convention.
}
\examples{
path_to_file <- '/path/to/data'
data <- importX(path_to_file)

path_to_file <- '/path/to/data.h5'
data <- importX(path = path_to_file, ESNG = T)
}
