% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate.r
\name{isolate}
\alias{isolate}
\title{isolate}
\usage{
isolate(sc_obj, max_depth = NULL, verbose = F)
}
\arguments{
\item{sc_obj}{an sc_obj with attached qc}

\item{max_depth}{Binary tree depth, default is 500.}

\item{verbose}{By default is set to False}
}
\value{
an sc_obj containing the identities of outliers in the metadata
}
\description{
To generate isolation forests isolate makes use of 'solitude'
https://CRAN.R-project.org/package=solitude
Komala Sheshachala Srikanth \link{aut, cre}, David Zimmermann \link{ctb}
}
\details{
outliers are detected using an isolation forest and their identities are
attached to the metadata of the sc_obj output.
}
\examples{
isolate(sc_obj)
}
