% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCobj.r
\name{newSC_obj,dgCMatrix-method}
\alias{newSC_obj,dgCMatrix-method}
\title{newSC_obj
generates a single cell object.}
\usage{
\S4method{newSC_obj}{dgCMatrix}(GeX_counts, HTO_counts)
}
\arguments{
\item{GeX_counts}{Takes a matrix of GeX counts in the form of a dgCMatrix.}

\item{HTO_counts}{Optionally accepts a matrix of HTO counts in the form of a dgCMatrix}
}
\value{
a single cell object with generated meta data.
}
\description{
newSC_obj
generates a single cell object.
}
\examples{
newSC_obj(GeX_counts)
##optionally##
newSC_obj(Gex_counts, HTO_counts)
}
