% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MULTIseq.r
\name{optimize_class}
\alias{optimize_class}
\title{optimize_class
takes a formatted table of HTO counts, and runs the MULTIseq classification
algorithm optimizing singlets:doublets:negatives}
\usage{
optimize_class(barTable)
}
\arguments{
\item{barTable}{Fomatted HTO count matrix from sc_obj slot HTO.dmplex slot HTO.matrix}
}
\value{
list of cell class calls
}
\description{
optimize_class
takes a formatted table of HTO counts, and runs the MULTIseq classification
algorithm optimizing singlets:doublets:negatives
}
