% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTOobj.r
\name{plotHashes}
\alias{plotHashes}
\title{Covenience function for visualizing output of barUMAP()}
\usage{
plotHashes(n.bar.table, UMAP.dims)
}
\arguments{
\item{bar.UMAP}{output matrix from barUMAP() function}

\item{plt}{logical, should plots be printed to std out. defaults to True.
Otherwise ggplot gobs are returned as a list.}
}
\value{
If plt is FALSE returns list of ggplot grobs otherwise no return value
}
\description{
Covenience function for visualizing output of barUMAP()
}
\examples{
plotHashes(bar.UMAP)
}
