% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.r
\name{rm_anomalies}
\alias{rm_anomalies}
\title{rm_anomalies}
\usage{
rm_anomalies(sc_obj)
}
\arguments{
\item{sc_obj}{an sc_obj with attached QC.}
}
\value{
sc_obj with anomalies removed from the active dataset
}
\description{
Convenience function which generate a new sc_obj assay containing the dataset
after anomalies have been removed. Raw data remains intact.
}
\examples{
  sce <- rm_anomalies(sce)
}
