% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCobj.r
\docType{class}
\name{SC_obj-class}
\alias{SC_obj-class}
\alias{SC_obj}
\title{SC_obj}
\description{
SC_obj
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{ANY.}

\item{\code{assay}}{list.}

\item{\code{meta.data}}{data.frame.}

\item{\code{graphs}}{list.}

\item{\code{ident}}{factor.}

\item{\code{est_doublet_rate}}{numeric.}
}}

