% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.r
\name{scatter}
\alias{scatter}
\title{scatter}
\usage{
scatter(data, x, y, color = NULL)
}
\arguments{
\item{data}{dataset to generate plot from}

\item{x}{data to map to x axis}

\item{y}{data to map to y axis}

\item{color}{color}
}
\value{
ggplot
}
\description{
Covenience function for generating scatter plots with ggplot2
}
